# back-end/docs/api/swagger.yaml
definitions:
  apperrors.AppError:
    properties:
      http_status:
        type: integer
      message:
        type: string
    type: object
...
  model.Deposit:
    properties:
      amount:
        type: string
      created_at:
        type: string
      external_id:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      status:
        $ref: '#/definitions/model.DepositStatus'
      transaction_id:
        type: string
      updated_at:
        type: string
      wallet_id:
        type: integer
    type: object
  model.DepositHistoryResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      project_id:
        type: integer
      project_name:
        type: string
      status:
        $ref: '#/definitions/model.DepositStatus'
      total_sum_sol:
        type: number
      transactions:
        items:
          $ref: '#/definitions/model.Transaction'
        type: array
    type: object
  model.DepositProcessResponse:
    properties:
      order_id:
        type: integer
      status:
        $ref: '#/definitions/model.DepositStatus'
    type: object
  model.DepositResponse:
    properties:
      amount:
        type: number
      fee:
        type: number
      order_id:
        type: integer
      status:
        $ref: '#/definitions/model.DepositStatus'
    type: object
  model.DepositSolanaReq:
    properties:
      account_id:
        type: integer
      max_amount:
        type: number
      min_amount:
        type: number
      project_id:
        type: integer
      quantity:
        type: integer
    required:
    - account_id
    - max_amount
    - min_amount
    - quantity
    type: object
  model.DepositStatus:
    enum:
    - PENDING
    - COMPLETED
    - FAILED
    - AWAITING_APPROVAL
    type: string
    x-enum-varnames:
    - DepositPending
    - DepositCompleted
    - DepositFailed
    - DepositAwaitingApproval
...
  model.Exchange:
    properties:
      id:
        example: 1234567890
        type: integer
      name:
        example: Binance
        type: string
    type: object
...
  model.Transaction:
    properties:
      balance_sol:
        type: number
      public_key:
        type: string
      status:
        $ref: '#/definitions/model.DepositStatus'
      sum_sol:
        type: number
      transaction_id:
        type: string
    type: object
...
host: https://api-stage.tsunammi.io/
info:
  contact: {}
  description: This is the swagger specification for Tsunammi API.
  title: Tsunammi API
  version: 0.0.7
paths:
...
  /exchanges:
    get:
      description: Returns all exchanges
      operationId: get-exchanges
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Exchange'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: List exchanges
      tags:
      - exchanges
...
  /wallets/solana/deposit:
    post:
      consumes:
      - application/json
      description: Creates a new Solana deposit request
      operationId: deposit-solana
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Deposit request payload
        in: body
        name: deposit
        required: true
        schema:
          $ref: '#/definitions/model.DepositSolanaReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.DepositResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Create Solana deposit
      tags:
      - deposits
  /wallets/solana/deposit/{id}:
    post:
      description: Returns the current status of a deposit by its ID
      operationId: get-deposit-status
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Deposit ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Deposit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Get deposit status
      tags:
      - deposits
  /wallets/solana/deposit/history:
    get:
      description: Returns the deposit history for the authenticated user.
      operationId: get-deposit-history
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DepositHistoryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Get deposit history
      tags:
      - deposits
  /wallets/solana/deposit/history/{id}:
    get:
      description: Returns the deposit history for a specific project ID, owned by
        the authenticated user.
      operationId: get-deposit-history-by-project-id
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DepositHistoryResponse'
        "400":
          description: Bad Request (invalid ID)
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Get deposit history by project ID
      tags:
      - deposits
  /wallets/solana/deposit/process/{id}:
    post:
      description: Processes a previously created deposit order by its ID.
      operationId: process-deposit-order
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Deposit Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/model.DepositProcessResponse'
        "400":
          description: Bad Request (invalid ID)
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Process a deposit order
      tags:
      - deposits
...
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer' followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"